[Refactor(christmas)] 패키지 변경
